python_example_email/requirements.txt
flask==2.0.1
werkzeug==2.0.3
python-dotenv==0.19.0
flask-cors==3.0.10 